Backend/service_orders/app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-multipart
PyJWT
passlib
